Extract generic trait from postgresql structs
In order to allow multiple database implementations,
`Connection` and `DbResult` have been promoted to Traits instead of structs

A new folder *postgresql* have been added with the relevant postgresql
implementation.
