+++ diesel/src/lib.rs
@@ -19,6 +19,7 @@ mod query_dsl;
 pub mod query_source;
 pub mod result;
 mod row;
+pub mod postgresql;
 
 pub mod helper_types {
     //! Provide helper types for concisely writing the return type of functions.
