+++ diesel/src/postgresql/mod.rs
@@ -1,0 +1,8 @@
+extern crate pq_sys;
+extern crate libc;
+
+pub mod db_result;
+pub mod connection;
+#[doc(hidden)]
+pub mod query_builder;
+
