+++ diesel/src/query_builder/mod.rs
@@ -1,6 +1,4 @@
 //! Contains traits responsible for the actual construction of SQL statements
-#[doc(hidden)]
-pub mod pg;
 pub mod debug;
 
 mod delete_statement;
