+++ diesel/src/row.rs
@@ -5,14 +5,14 @@ pub trait Row {
     fn next_is_null(&self, count: usize) -> bool;
 }
 
-pub struct DbRow<'a> {
-    db_result: &'a DbResult,
+pub struct DbRow<'a, R: 'a> {
+    db_result: &'a R,
     row_idx: usize,
     col_idx: usize,
 }
 
-impl<'a> DbRow<'a> {
-    pub fn new(db_result: &'a DbResult, row_idx: usize) -> Self {
+impl<'a, R: DbResult> DbRow<'a, R> {
+    pub fn new(db_result: &'a R, row_idx: usize) -> Self {
         DbRow {
             db_result: db_result,
             row_idx: row_idx,
@@ -21,7 +21,7 @@ impl<'a> DbRow<'a> {
     }
 }
 
-impl<'a> Row for DbRow<'a> {
+impl<'a, R: DbResult> Row for DbRow<'a, R> {
     fn take(&mut self) -> Option<&[u8]> {
         let current_idx = self.col_idx;
         self.col_idx += 1;
